requirements.txt
asgiref==3.2.10
Django==3.1.1
Faker==0.9.1
mixer==6.1.3
Pillow==7.2.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
text-unidecode==1.2
redis==3.5.3
celery==5.0.0
djangorestframework==3.12.1
